#394: Refactor the detail page to not use SearchKit when retrieving an individual study.
This involves creating a new action, updateStudy, that retrieves an individual study from Elasticsearch and performs all the required processing to gather all of the information needed to display the detail page. This makes the app easer to understand, as the behaviour of an external dependency is no longer relied upon.

This commit also includes the start of the work in order to have a language parameter in the URL. At the moment, it doesn't consistently work.